Refactored for HL7 server mode
