+++ dependency-reduced-pom.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>net.development.hl7messaging</groupId>
-  <artifactId>hl7messaging</artifactId>
+  <groupId>net.development.hl7messagingServer</groupId>
+  <artifactId>hl7messagingServer</artifactId>
   <name>JGroups :: Reliable Messaging</name>
   <version>1.0.0-SNAPSHOT</version>
-  <description>HL7 communication with JGroups</description>
-  <url>http://www.example.com</url>
+  <description>HL7 communication with JGroups - Server mode</description>
   <build>
     <plugins>
       <plugin>
@@ -23,7 +22,7 @@
         <configuration>
           <transformers>
             <transformer>
-              <mainClass>net.development.jgroupshl7.HL7StartInterface</mainClass>
+              <mainClass>net.development.jgroupshl7.server.HL7StartInterface</mainClass>
             </transformer>
           </transformers>
           <filters>
