+++ pom.xml
@@ -4,14 +4,14 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>net.development.hl7messaging</groupId>
-  <artifactId>hl7messaging</artifactId>
+  <groupId>net.development.hl7messagingServer</groupId>
+  <artifactId>hl7messagingServer</artifactId>
   <version>1.0.0-SNAPSHOT</version>
 
   <name>JGroups :: Reliable Messaging</name>
-  <description>HL7 communication with JGroups</description>
+  <description>HL7 communication with JGroups - Server mode</description>
   <!-- FIXME change it to the project's website -->
-  <url>http://www.example.com</url>
+  <!-- <url>http://www.example.com</url> -->
   <packaging>jar</packaging>
 
   <properties>
@@ -62,7 +62,7 @@
             <configuration>
                 <transformers>
                     <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                        <mainClass>net.development.jgroupshl7.HL7StartInterface</mainClass>
+                        <mainClass>net.development.jgroupshl7.server.HL7StartInterface</mainClass>
                     </transformer>
                 </transformers>
                 <filters>
