+++ src/main/java/net/development/jgroupshl7/server/HL7MessageSender.java
@@ -1,4 +1,4 @@
-package net.development.jgroupshl7;
+package net.development.jgroupshl7.server;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
